FEATURE:Change logic with google auth
